modified to accomodate all new requirements to merge to main
