+++ backend/src/types/track.ts
@@ -1,6 +1,8 @@
+// track.ts
 import { z } from 'zod';
 import { Album } from './album.ts';
 import { Artist } from './artist.ts';
+import { User } from './user.ts'; // Import User for PlaylistTrack
 
 export const Track = z.object({
     id: z.number(),
@@ -13,7 +15,19 @@ export const Track = z.object({
     popularity: z.number().int().min(0).max(100),
     previewUrl: z.string(),
     trackNumber: z.number(),
-    uri: z.string()
+    uri: z.string(),
 });
 
 export type TrackZ = z.input<typeof Track>;
+
+// PlaylistTrack definition moved here
+export const PlaylistTrack = z.object({
+    id: z.number(),
+    addedAt: z.date(),
+    addedBy: User,
+    track: Track,
+    // Removed playlist since it's not needed
+    // playlist: Playlist
+});
+
+export type PlaylistTrackZ = z.input<typeof PlaylistTrack>;
